functions/evaluate_answer/requirements.txt
google-cloud-bigquery
google-generativeai
functions-framework